+++ src/redturtle/volto/restapi/serializer/dxfields.py
@@ -8,6 +8,11 @@ from redturtle.volto.interfaces import IRedturtleVoltoLayer
 from zope.component import adapter
 from zope.component import getMultiAdapter
 from zope.schema.interfaces import ITextLine
+from plone.dexterity.interfaces import IDexterityContent
+from plone.restapi.interfaces import IFieldSerializer
+from zope.interface import implementer
+from zope.schema.interfaces import IDatetime
+from plone.app.dexterity.behaviors.metadata import IPublication
 
 import re
 
@@ -35,3 +40,27 @@ class TextLineFieldSerializer(DefaultFieldSerializer):
                 value = ref_obj.absolute_url()
 
         return json_compatible(value)
+
+
+@adapter(IDatetime, IDexterityContent, IRedturtleVoltoLayer)
+@implementer(IFieldSerializer)
+class DateTimeFieldSerializer:
+    def __init__(self, field, context, request):
+        self.context = context
+        self.request = request
+        self.field = field
+
+    def __call__(self):
+        return json_compatible(self.get_value())
+
+    def get_value(self, default=None):
+        value = getattr(
+            self.field.interface(self.context), self.field.__name__, default
+        )
+        if value and self.field.interface == IPublication:
+            # the patch: we want the dates with full tz infos
+            # default value is taken from
+            # plone.app.dexterity.behaviors.metadata.Publication that escape
+            # timezone
+            return getattr(self.context, self.field.__name__)()
+        return value